# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sharkie",
    "description": "Exchange rates API with historical data back to 2000",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://sharkie.uha.app",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/currencies": {
      "get": {
        "tags": [
          "currencies"
        ],
        "operationId": "get_currencies",
        "responses": {
          "200": {
            "description": "Available currency codes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrenciesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Service healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service degraded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/history": {
      "get": {
        "tags": [
          "history"
        ],
        "operationId": "get_history",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "Comma-separated dates (YYYY-MM-DD). Mutually exclusive with from/to.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Range start date (YYYY-MM-DD). Use with `to`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Range end date (YYYY-MM-DD). Use with `from`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currencies",
            "in": "query",
            "description": "Comma-separated currency codes to filter (e.g. \"EUR,GBP\").",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "base",
            "in": "query",
            "description": "Base currency for rate conversion. Defaults to USD.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max number of days to return (1-366). Defaults to 366.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of days to skip for pagination. Defaults to 0.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historical exchange rates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No data found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Response too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CurrenciesResponse": {
        "type": "object",
        "required": [
          "currencies"
        ],
        "properties": {
          "currencies": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Sorted list of available currency codes."
          }
        }
      },
      "DayRates": {
        "type": "object",
        "required": [
          "date",
          "rates"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date in YYYY-MM-DD format."
          },
          "rates": {
            "type": "object",
            "description": "Currency code → exchange rate relative to base.",
            "additionalProperties": {
              "type": "number",
              "format": "double"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "status",
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description."
          },
          "status": {
            "type": "string",
            "description": "Always \"error\"."
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status",
          "message",
          "db"
        ],
        "properties": {
          "db": {
            "type": "boolean",
            "description": "Whether the database is reachable."
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "\"ok\" or \"degraded\"."
          }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "required": [
          "base",
          "total",
          "data"
        ],
        "properties": {
          "base": {
            "type": "string",
            "description": "Base currency used for rate conversion."
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayRates"
            },
            "description": "Paginated list of daily rates."
          },
          "total": {
            "type": "integer",
            "description": "Total number of days matching the query (before pagination).",
            "minimum": 0
          }
        }
      }
    }
  }
}